Take out MP and CR review methods form submission_review
